+++ azure/storage/queue/models.py
@@ -12,72 +12,72 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #--------------------------------------------------------------------------
-from .._common_models import (
-    WindowsAzureData,
-    _list_of,
-)
-from ..models import (
-    EnumResultsBase,
-)
 
+class Queue(object):
 
-class QueueEnumResults(EnumResultsBase):
-
-    ''' Queue list'''
-
-    def __init__(self):
-        EnumResultsBase.__init__(self)
-        self.queues = _list_of(Queue)
-
-    def __iter__(self):
-        return iter(self.queues)
-
-    def __len__(self):
-        return len(self.queues)
-
-    def __getitem__(self, index):
-        return self.queues[index]
-
-
-class Queue(WindowsAzureData):
-
-    ''' Queue class '''
+    ''' 
+    Queue class.
+     
+    :ivar name: 
+        The name of the queue.
+    :vartype name: str
+    :ivar metadata: 
+        A dict containing name-value pairs associated with the queue as metadata.
+        This var is set to None unless the include=metadata param was included 
+        for the list queues operation. If this parameter was specified but the 
+        queue has no metadata, metadata will be set to an empty dictionary.
+    :vartype metadata: dict
+    '''
 
     def __init__(self):
-        self.name = u''
-        self.url = u''
-        self.metadata = {}
+        self.name = None
+        self.metadata = None
 
 
-class QueueMessagesList(WindowsAzureData):
+class QueueMessage(object):
 
-    ''' Queue message list. '''
+    ''' 
+    Queue message class. 
+
+    :ivar message_id: 
+        A GUID value assigned to the message by the Queue service that 
+        identifies the message in the queue. This value may be used together 
+        with the value of pop_receipt to delete a message from the queue after 
+        it has been retrieved with the get messages operation. 
+    :vartype message_id: str
+    :ivar insertion_time: 
+        A UTC date value representing the time the messages was inserted.
+    :vartype insertion_time: date
+    :ivar expiration_time: 
+        A UTC date value representing the time the message expires.
+    :vartype expiration_time: date
+    :ivar dequeue_count: 
+        Begins with a value of 1 the first time the message is dequeued. This 
+        value is incremented each time the message is subsequently dequeued.
+    :vartype dequeue_count: int
+    :ivar message_text: 
+        A UTC date value representing the time the messages was inserted.
+    :vartype message_text: date
+    :ivar pop_receipt: 
+        A receipt str which can be used together with the message_id element to 
+        delete a message from the queue after it has been retrieved with the get 
+        messages operation. Only returned by get messages operations. Set to 
+        None for peek messages.
+    :vartype pop_receipt: str
+    :ivar time_next_visible: 
+        A UTC date value representing the time the message will next be visible. 
+        Only returned by get messages operations. Set to None for peek messages.
+    :vartype time_next_visible: date
+    '''
 
     def __init__(self):
-        self.queue_messages = _list_of(QueueMessage)
-
-    def __iter__(self):
-        return iter(self.queue_messages)
-
-    def __len__(self):
-        return len(self.queue_messages)
-
-    def __getitem__(self, index):
-        return self.queue_messages[index]
-
-
-class QueueMessage(WindowsAzureData):
-
-    ''' Queue message class. '''
-
-    def __init__(self):
-        self.message_id = u''
-        self.insertion_time = u''
-        self.expiration_time = u''
-        self.pop_receipt = u''
-        self.time_next_visible = u''
-        self.dequeue_count = u''
-        self.message_text = u''
+        self.message_id = None
+        self.insertion_time = None
+        self.expiration_time = None
+        self.dequeue_count = None
+        self.message_text = None
+        self.pop_receipt = None
+        self.time_next_visible = None
 
 
 class QueueSharedAccessPermissions(object):